fast-url-handler/app/requirements.txt
flask==3.0.0
redis==5.0.1
cassandra-driver==3.28.0